+++ packages/plugin-msw/src/generators/mswGenerator.tsx
@@ -12,6 +12,7 @@ export const mswGenerator = createReactGenerator<PluginMsw>({
   name: 'msw',
   Operation({ operation }) {
     const {
+      pluginManager,
       plugin: {
         options: { output, parser, baseURL },
       },
@@ -35,7 +36,13 @@ export const mswGenerator = createReactGenerator<PluginMsw>({
     }
 
     return (
-      <File baseName={mock.file.baseName} path={mock.file.path} meta={mock.file.meta} banner={getBanner({ oas, output })} footer={getFooter({ oas, output })}>
+      <File
+        baseName={mock.file.baseName}
+        path={mock.file.path}
+        meta={mock.file.meta}
+        banner={getBanner({ oas, output, config: pluginManager.config })}
+        footer={getFooter({ oas, output })}
+      >
         <File.Import name={['http']} path="msw" />
         <File.Import name={['ResponseResolver']} isTypeOnly path="msw" />
         <File.Import name={[type.schemas.response.name]} path={type.file.path} root={mock.file.path} isTypeOnly />