+++ packages/plugin-react-query/src/generators/infiniteQueryGenerator.tsx
@@ -79,7 +79,7 @@ export const infiniteQueryGenerator = createReactGenerator<PluginReactQuery>({
         baseName={query.file.baseName}
         path={query.file.path}
         meta={query.file.meta}
-        banner={getBanner({ oas, output })}
+        banner={getBanner({ oas, output, config: pluginManager.config })}
         footer={getFooter({ oas, output })}
       >
         {options.parser === 'zod' && (