+++ packages/plugin-svelte-query/src/generators/queryGenerator.tsx
@@ -76,7 +76,7 @@ export const queryGenerator = createReactGenerator<PluginSvelteQuery>({
         baseName={query.file.baseName}
         path={query.file.path}
         meta={query.file.meta}
-        banner={getBanner({ oas, output })}
+        banner={getBanner({ oas, output, config: pluginManager.config })}
         footer={getFooter({ oas, output })}
       >
         {options.parser === 'zod' && (