+++ packages/plugin-swr/src/generators/mutationGenerator.tsx
@@ -74,7 +74,7 @@ export const mutationGenerator = createReactGenerator<PluginSwr>({
         baseName={mutation.file.baseName}
         path={mutation.file.path}
         meta={mutation.file.meta}
-        banner={getBanner({ oas, output })}
+        banner={getBanner({ oas, output, config: pluginManager.config })}
         footer={getFooter({ oas, output })}
       >
         {options.parser === 'zod' && (