+++ packages/kubb/CHANGELOG.md
@@ -1,5 +1,29 @@
 # kubb
 
+## 3.0.59
+
+### Patch Changes
+
+- Updated dependencies [[`f8cfede`](https://github.com/kubb-labs/kubb/commit/f8cfedee78bb3ff81ba0dcc8e68dc9172913dbe0)]:
+  - @kubb/core@3.9.0
+  - @kubb/cli@3.9.0
+
+## 3.0.58
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/cli@3.8.1
+  - @kubb/core@3.8.1
+
+## 3.0.57
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @kubb/cli@3.8.0
+  - @kubb/core@3.8.0
+
 ## 3.0.56
 
 ### Patch Changes